backend: move chart generation from handlers to backend
First step in a small series of refactoring improvements.

This commit:

Handlers just handle http input/output, the main functionality should
not be there. This makes it consistent with most other handlers.

Next commits: split into multiple smaller functions and get rid of the
interface{} return type.
