+++ backend/handlers/handlers.go
@@ -21,12 +21,10 @@ import (
 	"encoding/json"
 	"fmt"
 	"io"
-	"math/big"
 	"net/http"
 	"os"
 	"path/filepath"
 	"runtime/debug"
-	"sort"
 	"strconv"
 	"strings"
 	"time"
@@ -35,13 +33,11 @@ import (
 	"github.com/btcsuite/btcutil/hdkeychain"
 	"github.com/digitalbitbox/bitbox-wallet-app/backend"
 	"github.com/digitalbitbox/bitbox-wallet-app/backend/accounts"
-	"github.com/digitalbitbox/bitbox-wallet-app/backend/accounts/errors"
 	"github.com/digitalbitbox/bitbox-wallet-app/backend/banners"
 	"github.com/digitalbitbox/bitbox-wallet-app/backend/bitboxbase"
 	baseHandlers "github.com/digitalbitbox/bitbox-wallet-app/backend/bitboxbase/handlers"
 	"github.com/digitalbitbox/bitbox-wallet-app/backend/coins/btc"
 	accountHandlers "github.com/digitalbitbox/bitbox-wallet-app/backend/coins/btc/handlers"
-	"github.com/digitalbitbox/bitbox-wallet-app/backend/coins/coin"
 	coinpkg "github.com/digitalbitbox/bitbox-wallet-app/backend/coins/coin"
 	"github.com/digitalbitbox/bitbox-wallet-app/backend/coins/eth"
 	"github.com/digitalbitbox/bitbox-wallet-app/backend/config"
@@ -112,6 +108,7 @@ type Backend interface {
 	CheckForUpdateIgnoringErrors() *backend.UpdateFile
 	Banners() *banners.Banners
 	Environment() backend.Environment
+	ChartData() (interface{}, error)
 }
 
 // Handlers provides a web api to the backend.
@@ -782,237 +779,8 @@ func (handlers *Handlers) apiMiddleware(devMode bool, h func(*http.Request) (int
 		writeJSON(w, value)
 	})
 }
-
-func (handlers *Handlers) allCoinCodes() []string {
-	allCoinCodes := []string{}
-	for _, account := range handlers.backend.Accounts() {
-		if account.FatalError() {
-			continue
-		}
-		allCoinCodes = append(allCoinCodes, string(account.Coin().Code()))
-	}
-	return allCoinCodes
-}
-
 func (handlers *Handlers) getAccountSummary(_ *http.Request) (interface{}, error) {
-	type chartEntry struct {
-		Time  int64   `json:"time"`
-		Value float64 `json:"value"`
-	}
-
-	// If true, we are missing headers or historical conversion rates necessary to compute the chart
-	// data,
-	chartDataMissing := false
-
-	// key: unix timestamp.
-	chartEntriesDaily := map[int64]chartEntry{}
-	chartEntriesHourly := map[int64]chartEntry{}
-
-	fiat := handlers.backend.Config().AppConfig().Backend.MainFiat
-	// Chart data until this point in time.
-	until := handlers.backend.RatesUpdater().HistoryLatestTimestampAll(handlers.allCoinCodes(), fiat)
-	if until.IsZero() {
-		chartDataMissing = true
-		handlers.log.Info("ChartDataMissing, until is zero")
-	}
-	isUpToDate := time.Since(until) < 2*time.Hour
-	handlers.log.Infof("Chart/isUptodate: %v", isUpToDate)
-
-	currentTotal := new(big.Rat)
-	currentTotalMissing := false
-	// Total number of transactions across all active accounts.
-	totalNumberOfTransactions := 0
-	for _, account := range handlers.backend.Accounts() {
-		if account.FatalError() {
-			continue
-		}
-		err := account.Initialize()
-		if err != nil {
-			return nil, err
-		}
-		balance, err := account.Balance()
-		if err != nil {
-			return nil, err
-		}
-		txs, err := account.Transactions()
-		if err != nil {
-			return nil, err
-		}
-		totalNumberOfTransactions += len(txs)
-
-		// e.g. 1e8 for Bitcoin/Litecoin, 1e18 for Ethereum, etc. Used to convert from the smallest
-		// unit to the standard unit (BTC, LTC; ETH, etc.).
-		coinDecimals := new(big.Int).Exp(
-			big.NewInt(10),
-			big.NewInt(int64(account.Coin().Decimals(false))),
-			nil,
-		)
-
-		// HACK: We still use the latest prices from CryptoCompare for the account fiat balances
-		// above (displayed in the summary table). Those might deviate from the latest historical
-		// prices from coingecko, which results in different total balances in the chart and the
-		// summary table.
-		//
-		// As a temporary workaround, until we use only one source for all prices, we manually
-		// compute the total based on the latest rates from CryptoCompare.
-		price, err := handlers.backend.RatesUpdater().LastForPair(string(account.Coin().Code()), fiat)
-		if err != nil {
-			currentTotalMissing = true
-			handlers.log.
-				WithField("coin", account.Coin().Code()).WithError(err).Info("currentTotalMissing")
-		}
-		fiatValue := new(big.Rat).Mul(
-			new(big.Rat).SetFrac(
-				balance.Available().BigInt(),
-				coinDecimals,
-			),
-			new(big.Rat).SetFloat64(price),
-		)
-		currentTotal.Add(currentTotal, fiatValue)
-
-		// Below here, only chart data is being computed.
-		if chartDataMissing {
-			continue
-		}
-
-		// Time from which the chart turns from daily points to hourly points.
-		hourlyFrom := time.Now().AddDate(0, 0, -7).Truncate(24 * time.Hour)
-
-		earliestPriceAvailable := handlers.backend.RatesUpdater().HistoryEarliestTimestamp(
-			string(account.Coin().Code()),
-			fiat)
-
-		earliestTxTime, err := txs.EarliestTime()
-		if errp.Cause(err) == errors.ErrNotAvailable {
-			handlers.log.WithField("coin", account.Coin().Code()).Info("ChartDataMissing/earliestTxtime")
-			chartDataMissing = true
-			continue
-		}
-		if err != nil {
-			return nil, err
-		}
-
-		if earliestTxTime.IsZero() {
-			// Ignore the chart for this account, there is no timed transaction.
-			continue
-		}
-		if earliestPriceAvailable.IsZero() || earliestTxTime.Before(earliestPriceAvailable) {
-			chartDataMissing = true
-			handlers.log.
-				WithField("coin", account.Coin().Code()).
-				WithField("earliestTxTime", earliestTxTime).
-				WithField("earliestPriceAvailable", earliestPriceAvailable).
-				Info("ChartDataMissing")
-			continue
-		}
-
-		timeseriesDaily, err := txs.Timeseries(
-			earliestTxTime.Truncate(24*time.Hour),
-			until,
-			24*time.Hour,
-		)
-		if errp.Cause(err) == errors.ErrNotAvailable {
-			handlers.log.WithField("coin", account.Coin().Code()).Info("ChartDataMissing")
-			chartDataMissing = true
-			continue
-		}
-		if err != nil {
-			return nil, err
-		}
-		timeseriesHourly, err := txs.Timeseries(
-			hourlyFrom,
-			until,
-			time.Hour,
-		)
-		if errp.Cause(err) == errors.ErrNotAvailable {
-			handlers.log.WithField("coin", account.Coin().Code()).Info("ChartDataMissing")
-			chartDataMissing = true
-			continue
-		}
-		if err != nil {
-			return nil, err
-		}
-
-		addChartData := func(coinCode coin.Code, timeseries []accounts.TimeseriesEntry, chartEntries map[int64]chartEntry) {
-			for _, e := range timeseries {
-				price := handlers.backend.RatesUpdater().PriceAt(
-					string(coinCode),
-					fiat,
-					e.Time)
-				timestamp := e.Time.Unix()
-				chartEntry := chartEntries[timestamp]
-
-				chartEntry.Time = timestamp
-				fiatValue, _ := new(big.Rat).Mul(
-					new(big.Rat).SetFrac(
-						e.Value.BigInt(),
-						coinDecimals,
-					),
-					new(big.Rat).SetFloat64(price),
-				).Float64()
-				chartEntry.Value += fiatValue
-				chartEntries[timestamp] = chartEntry
-			}
-		}
-
-		addChartData(account.Coin().Code(), timeseriesDaily, chartEntriesDaily)
-		addChartData(account.Coin().Code(), timeseriesHourly, chartEntriesHourly)
-
-	}
-
-	toSortedSlice := func(s map[int64]chartEntry) []chartEntry {
-		result := make([]chartEntry, len(s))
-		i := 0
-		for _, entry := range s {
-			result[i] = entry
-			i++
-		}
-		sort.Slice(result, func(i, j int) bool { return result[i].Time < result[j].Time })
-
-		// Manually add the last point with the current total, to make the last point match.
-		// The last point might not match the account total otherwise because:
-		// 1) unconfirmed tx are not in the timeseries
-		// 2) coingecko might not have rates yet up until after all transactions, so they'd also be
-		// missing form the timeseries (`until` is up to 2h in the past).
-		if isUpToDate && !currentTotalMissing {
-			total, _ := currentTotal.Float64()
-			result = append(result, chartEntry{
-				Time:  time.Now().Unix(),
-				Value: total,
-			})
-		}
-
-		// Truncate leading zeroes.
-		for i, e := range result {
-			if e.Value > 0 {
-				return result[i:]
-			}
-		}
-		// Everything was zeroes.
-		return []chartEntry{}
-	}
-
-	// Even if we are still gathering data (exchange rates, block headers), we know the result
-	// already if there are no transactions. This avoids showing the user a message that we are
-	// gathering data, only to show nothing in the end.
-	if chartDataMissing && totalNumberOfTransactions == 0 {
-		handlers.log.Info("ChartDataMissing forced to false")
-		chartDataMissing = false
-	}
-
-	var chartTotal *float64
-	if !currentTotalMissing {
-		tot, _ := currentTotal.Float64()
-		chartTotal = &tot
-	}
-	return map[string]interface{}{
-		"chartDataMissing": chartDataMissing,
-		"chartDataDaily":   toSortedSlice(chartEntriesDaily),
-		"chartDataHourly":  toSortedSlice(chartEntriesHourly),
-		"chartFiat":        fiat,
-		"chartTotal":       chartTotal,
-		"chartIsUpToDate":  isUpToDate, // only valid is chartDataMissing is false
-	}, nil
+	return handlers.backend.ChartData()
 }
 
 func (handlers *Handlers) postExportAccountSummary(_ *http.Request) (interface{}, error) {
